remove debug breakpoints, format, add readme, tweak cuda jit code
