+++ threedgrt_tracer/tracer.py
@@ -231,7 +231,6 @@ class Tracer:
                 self.conf.render.min_transmittance,
             )
 
-            import pdb; pdb.set_trace()
             if self.frame_timer is not None:
                 self.frame_timer.end()
