+++ threedgrut/model/model.py
@@ -490,8 +490,10 @@ class MixtureOfGaussians(torch.nn.Module):
 
         if self.conf.optimizer.type == "adam":
             self.optimizer = torch.optim.Adam(params, lr=self.conf.optimizer.lr, eps=self.conf.optimizer.eps)
+            logger.info("🔆 Using Adam optimizer")
         elif self.conf.optimizer.type == "selective_adam":
             self.optimizer = SelectiveAdam(params, lr=self.conf.optimizer.lr, eps=self.conf.optimizer.eps)
+            logger.info("🔆 Using Selective Adam optimizer")
         else:
             raise ValueError(f"Unknown optimizer type: {self.conf.optimizer.type}")
 
@@ -573,8 +575,6 @@ class MixtureOfGaussians(torch.nn.Module):
         Returns:
             A dictionary containing the output of the model
         """
-        import pdb; pdb.set_trace()
-
         return self.renderer.render(self, batch, train, frame_id)
 
     def trace(self, rays_o, rays_d, T_to_world=None):
